# api/docs/swagger.yaml
definitions:
  handler.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: integer
    type: object
...
  models.CreateStore:
    properties:
      city:
        type: string
      email:
        type: string
      phone:
        type: string
      state:
        type: string
      store_name:
        type: string
      street:
        type: string
      zip_code:
        type: string
    type: object
...
  models.PatchRequest:
    properties:
      fields:
        additionalProperties: true
        type: object
      id:
        type: integer
    type: object
...
  models.UpdateStore:
    properties:
      city:
        type: string
      email:
        type: string
      phone:
        type: string
      state:
        type: string
      store_id:
        type: integer
      store_name:
        type: string
      street:
        type: string
      zip_code:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /store:
    get:
      consumes:
      - application/json
      description: Get List Store
      operationId: get_list_store
      parameters:
      - description: offset
        in: query
        name: offset
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get List Store
      tags:
      - Store
    post:
      consumes:
      - application/json
      description: Create Store
      operationId: create_store
      parameters:
      - description: CreateStoreRequest
        in: body
        name: store
        required: true
        schema:
          $ref: '#/definitions/models.CreateStore'
      produces:
      - application/json
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create Store
      tags:
      - Store
  /store/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Store
      operationId: get_by_id_store
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete Store
      tags:
      - Store
    get:
      consumes:
      - application/json
      description: Get By ID Store
      operationId: get_by_id_store
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get By ID Store
      tags:
      - Store
    patch:
      consumes:
      - application/json
      description: Update Patch Store
      operationId: updat_patch_store
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdatPatchStoreRequest
        in: body
        name: store
        required: true
        schema:
          $ref: '#/definitions/models.PatchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Patch Store
      tags:
      - Store
    put:
      consumes:
      - application/json
      description: Update Put Store
      operationId: updat_patch_store
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdateStore
        in: body
        name: store
        required: true
        schema:
          $ref: '#/definitions/models.UpdateStore'
      produces:
      - application/json
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Put Store
      tags:
      - Store
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"